requirements.txt
numpy
tensorflow-gpu >= 0.12
joblib
nltk >= 3.2.2
tqdm
pyprind
